+++ src/components/TableColumns.tsx
@@ -1,0 +1,2 @@
+export const mdBreakpoint = ["md" as const];
+export const smBreakPoint = ["sm" as const];
